update FSDP test. add autocast support

+++ torchao/prototype/quantized_training/int8.py
@@ -111,7 +111,7 @@ class Int8QuantizedTrainingLinearWeight(TorchAOBaseTensor):
         out: Optional[Tensor] = None,
     ):
         int_data, scale = all_gather_outputs
-        return Int8QuantizedTrainingLinearWeight(int_data, scale.to(param_dtype)), all_gather_outputs
+        return Int8QuantizedTrainingLinearWeight(int_data, scale), all_gather_outputs
 
 
 class _Int8WeightOnlyLinear(torch.autograd.Function):
